requirements.txt
# Telegram Bot Dependencies
aiogram>=3.4.0
aiohttp>=3.9.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
babel>=2.14.0
python-dotenv>=1.0.0
tonutils>=0.4.0
